Monster prevent investigators for investigating.

+++ app/models/ga/close_portal.rb
@@ -3,19 +3,23 @@ class Ga::ClosePortal < Ga::AskPeople
   def set_action_result
     clues_count = investigator.clues
 
-    if clues_count >= 3
-      portal = Portal.find_by_game_board_id_and_place_id( investigator.game_board_id, location_id )
-      if portal
-        ActiveRecord::Base.transaction do
-          portal.destroy
-          investigator.update_attribute( :clues, investigator[ :clues ] - 2 )
+    unless check_for_monster
+      if clues_count >= 3
+        portal = Portal.find_by_game_board_id_and_place_id( investigator.game_board_id, location_id )
+        if portal
+          ActiveRecord::Base.transaction do
+            portal.destroy
+            investigator.update_attribute( :clues, investigator[ :clues ] - 2 )
+          end
+          @action_result = :portal_closed
+        else
+          @action_result = :no_portal
         end
-        @action_result = :portal_closed
       else
-        @action_result = :no_portal
+        @action_result = :not_enough_clues
       end
     else
-      @action_result = :not_enough_clues
+      @action_result = :blocked_by_monster
     end
   end
